Several changes. Updated OPUtil functions to all have the same order of required variables: Concentration, Pressure, Temperature. Not all those variables are used for every function, but it helps for consistence. Updated example/SEOP/case.sif file to reflect this.
Updated all function (OPUtil, SESolver, and OPSolver) so that all numerical constants are of the form xx.0D0. This should ensure floating point math is done for all calculations. I was having trouble with some calculations being done with integer math...

Updated caculateevaprate so that it looks for "alpha" and "evap rate mass" parameters in the BC rather than in the "Materials" section of the SIF. The Materials section is hard to access when the function is calculating boundary conditions, so it is both easier (and conceptually makes more sense) to put that information in the BC where the function is called.