Lots of important changes in this commit. First, I've added code that gets the reference temperature from the material section. This is used when the Perfect Gas compressiblity model is used. It turns out, it is really hard for Elmer's Naveir Stokes solver to do pressure boundary conditions. So, Perfect Pressure compressiblity model solves the problem with the solved pressure solved for a round zero. The reference pressure is = to 0 pressure in the solver. So the solver solves for variations in that reference pressure. I need the reference pressure in order to calculate the absolute pressure for my spin-exchange parameters and thermo-fluid parameters.
This caused a big problem in the CalcualteDecayRate and CalcualteXenonDiffusion functions. You see, CalculateDecayRate calls CalculateXenonDiffusion because I need to know the diffusion rate in order to calculate the T1 "mass transfer coefficient". But this T1 decay rate is only calculated on the boundary, and the material section is not available on the boundary. So it throws an error. And really, I don't want the reference pressure for this calculation because I pass the pressure through Calculate DecayRate.

Right now, I'm using this stupid method. The element number "n" which the Elmer solver needs to use the user-defined functions are always positive. So, in the CalculateDecayRate function, I just pass a -1 through this variable and then have some logic that prevents the solver from looking for the reference pressure.

I think I can actually think of a better way to do this, and all implement it later this week, but I'll push this now. The better way to implement it is to have a function that is called by both CalculateXenonDiffusion and CalculateDecayRate, but doesn't interface with the rest of the Elmer program. This would actually solve a lot of problems and make it easy to implement a solver for Rb diffusion as well.

I also fixed an issue with the order of varaibles for the CalculateSpinDestrutionRate and CalcualteRbPol sovlers. They had incorrectly ordered the variable "concentration, Temperature, Pressure" when it should have been "concentration, Pressure, Temperature". This was causing problems with the laserheat function which calls both of those functions and uses the standard ordering.

I fixed this is the case.sif as well.

Also fixed some stuff in the interface. Just some variable declarations. However, I don't think this was actually a problem because the program seemed to be running fine with them being wrong.

The program compiles, and I've tested in in SEOP_2D, and get the same answers as previously.