Added checks to make sure both arguments for OPSolver, OPUtil, and SESolver were within bounds. Mainly, this meant checking the argument were greater than 0, and in some cases were bound between 0 and 1.
Also added checks to OPUtil functions to assure that output of functions was rational. Again, this mainly meant checking the output was greater than 0, and in some cases, was bounded between 0 and 1.

Finally, removed old code that was intended for a unit change for Rb number density. Would have converted kg/m^2 to /m^2. This is not necessary in the new scheme as plain number density will be used.